+++ src/object/assign/method.ts
@@ -1,0 +1,3 @@
+const method = (obj: object, ...sources: object[]): object => Object.assign(obj, ...sources);
+
+export = method;