+++ src/object/keys/method.ts
@@ -1,0 +1,3 @@
+const method = (obj: object): string[] => Object.keys(obj);
+
+export = method;