+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [v0.12.0](https://github.com/ardalanamini/prototyped.js/releases/tag/v0.12.0) *(2018-07-26)*
+
+**Implemented enhancements:**
+
+- `Object.prototype`
+  - Added `$assign`
+  - Added `$defaults`
+  - Added `$keys`
+
 ## [v0.11.0](https://github.com/ardalanamini/prototyped.js/releases/tag/v0.11.0) *(2018-07-16)*
 
 **Implemented enhancements:**